+++ js.bolt.surge/src/errors/index.js
@@ -1,6 +1,0 @@
-import {
-  EnvironmentVariableInvalidError,
-  EnvironmentVariableMissingError,
-} from "./errors.js";
-
-export { EnvironmentVariableInvalidError, EnvironmentVariableMissingError };